+++ tools/sdk-ai-bots/azure-sdk-qa-bot/src/input/PromptGeneratorV2.ts
